# backend/swagger.yaml
openapi: 3.0.0
info:
  title: ProfSale API
  description: API for ProfSale - Point of Sale and Business Management System for SMBs
  version: 1.0.0
  contact:
    name: ProfSale Support
    email: support@profsale.com
  license:
    name: MIT

servers:
  - url: https://profsale.dangotechconcepts.com/api/v1
    description: Production Server
  - url: http://localhost:5000/api/v1
    description: Development Server

tags:
  - name: Authentication
    description: User authentication and authorization
  - name: Products
    description: Product management
  - name: Sales
    description: Sales and POS operations
  - name: Customers
    description: Customer management
  - name: Dashboard
    description: Dashboard analytics and metrics
  - name: Business
    description: Business settings and configuration
  - name: Expenses
    description: Expense tracking
  - name: Reports
    description: Business reports
  - name: Staff
    description: Staff management
  - name: Subscriptions
    description: Subscription management

paths:
  /auth/register:
    post:
      tags:
        - Authentication
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - phone
                - email
                - firstName
                - lastName
                - businessName
                - businessType
                - password
              properties:
                phone:
                  type: string
                  example: "+256700123456"
                email:
                  type: string
                  format: email
                  example: "user@example.com"
                firstName:
                  type: string
                  example: "John"
                lastName:
                  type: string
                  example: "Doe"
                businessName:
                  type: string
                  example: "My Business"
                businessType:
                  type: string
                  enum: [retail, wholesale, service, restaurant]
                password:
                  type: string
                  format: password
                  minLength: 6
      responses:
        '200':
          description: User registered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      user:
                        type: object
                      token:
                        type: string
        '400':
          description: Validation error

  /auth/login:
    post:
      tags:
        - Authentication
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - login
                - password
              properties:
                login:
                  type: string
                  description: Phone or email
                password:
                  type: string
                  format: password
      responses:
        '200':
          description: Login successful
        '401':
          description: Invalid credentials

  /products:
    get:
      tags:
        - Products
      summary: Get all products
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: category
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Products retrieved successfully
      security:
        - bearerAuth: []

    post:
      tags:
        - Products
      summary: Create a new product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - buyingPrice
                - sellingPrice
              properties:
                name:
                  type: string
                category:
                  type: string
                buyingPrice:
                  type: number
                sellingPrice:
                  type: number
                currentStock:
                  type: integer
                minStockLevel:
                  type: integer
                unit:
                  type: string
      responses:
        '200':
          description: Product created successfully
      security:
        - bearerAuth: []

  /products/{id}:
    get:
      tags:
        - Products
      summary: Get product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Product retrieved successfully
      security:
        - bearerAuth: []

    put:
      tags:
        - Products
      summary: Update product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Product updated successfully
      security:
        - bearerAuth: []

    delete:
      tags:
        - Products
      summary: Delete product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Product deleted successfully
      security:
        - bearerAuth: []

  /sales:
    post:
      tags:
        - Sales
      summary: Create a new sale
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - items
                - totalAmount
              properties:
                customerId:
                  type: integer
                customerName:
                  type: string
                customerPhone:
                  type: string
                items:
                  type: array
                  items:
                    type: object
                    properties:
                      productId:
                        type: integer
                      quantity:
                        type: integer
                      unitPrice:
                        type: number
                subtotal:
                  type: number
                totalAmount:
                  type: number
                discountAmount:
                  type: number
                taxAmount:
                  type: number
                paymentMethod:
                  type: string
                  enum: [cash, card, mobile_money, bank_transfer]
      responses:
        '200':
          description: Sale created successfully
      security:
        - bearerAuth: []

    get:
      tags:
        - Sales
      summary: Get all sales
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: startDate
          in: query
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Sales retrieved successfully
      security:
        - bearerAuth: []

  /dashboard/overview:
    get:
      tags:
        - Dashboard
      summary: Get dashboard overview
      responses:
        '200':
          description: Dashboard data retrieved successfully
      security:
        - bearerAuth: []

  /business/profile:
    get:
      tags:
        - Business
      summary: Get business profile
      responses:
        '200':
          description: Business profile retrieved successfully
      security:
        - bearerAuth: []

    put:
      tags:
        - Business
      summary: Update business profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Business profile updated successfully
      security:
        - bearerAuth: []

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoint

  schemas:
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
        errors:
          type: array
          items:
            type: object
        timestamp:
          type: string
          format: date-time